+++ crates/rome_cli/src/cli_options.rs
@@ -1,6 +1,7 @@
 use bpaf::Bpaf;
 use std::str::FromStr;
 
+/// Global options applied to all commands
 #[derive(Debug, Clone, Bpaf)]
 pub struct CliOptions {
     /// Set the formatting mode for markup: "off" prints everything as plain text, "force" forces the formatting of markup using ANSI even if the console output is determined to be incompatible